requirements.txt
# Core
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Geospatial
rioxarray>=0.15.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0

# Satellite data
planetary-computer>=1.0.0
pystac-client>=0.7.0

# Weather data
herbie-data>=2024.1.0
xarray>=2024.1.0
cfgrib>=0.9.10

# ML
xgboost>=2.0.0
lightgbm>=4.0.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Utilities
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
